feat: updated build file
