+++ service/service-samples/sample-service-reactive-grpc-hello/src/main/java/io/americanexpress/service/hello/reactive/grpc/HelloApplication.java
@@ -11,6 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
+
 package io.americanexpress.service.hello.reactive.grpc;
 
 import org.springframework.boot.SpringApplication;
